# backend/api/docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateUser:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
    type: object
  models.LoginUser:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.Message:
    properties:
      _id:
        type: string
      content:
        type: string
      createdAt:
        type: string
      receiver:
        type: string
      sender:
        type: string
    type: object
...
  models.SendMessageM:
    properties:
      content:
        minLength: 1
        type: string
      receiver:
        type: string
      sender:
        type: string
    required:
    - content
    - receiver
    - sender
    type: object
...
  models.UserModel:
    properties:
      _id:
        type: string
      bio:
        type: string
      email:
        type: string
      followers:
        items:
          type: string
        type: array
      following:
        items:
          type: string
        type: array
      imageUrl:
        type: string
      name:
        type: string
    required:
    - email
    type: object
host: localhost:5000
info:
  contact: {}
  description: this is  Swagger docs for Social chat API created with golang and fiber
  title: Fiber Golang Rest API
  version: "1.0"
paths:
  /chat/get-user-unreadmsg:
    get:
      consumes:
      - application/json
      description: Get unread message count & records for user
      parameters:
      - description: user id
        in: query
        name: userid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get unread message count & records for user
      tags:
      - Chat
  /chat/getmsgsbynums:
    get:
      consumes:
      - application/json
      description: Get messages between two users by pagination
      parameters:
      - description: Starting page num
        in: query
        name: from
        required: true
        type: integer
      - description: first user id
        in: query
        name: firstuid
        required: true
        type: string
      - description: second user id
        in: query
        name: seconduid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Message'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: get messages by pagination
      tags:
      - Chat
  /chat/mark-msg-asreaded:
    get:
      consumes:
      - application/json
      description: mark messages as read for user, updates the record to isReadMessage=true,
        count=0
      parameters:
      - description: main user id
        in: query
        name: mainuid
        required: true
        type: string
      - description: other user id
        in: query
        name: otheruid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: mark messages as read for user
      tags:
      - Chat
  /chat/sendmessage:
    post:
      consumes:
      - application/json
      description: SendMessage from one user to another
      parameters:
      - description: user SendMessage details
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.SendMessageM'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: send message to friend user
      tags:
      - Chat
...
  /user/refresh:
    get:
      consumes:
      - application/json
      description: refresh user data and issue a new token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: refresh user data and token
      tags:
      - Authentication
  /user/signin:
    post:
      consumes:
      - application/json
      description: Login a user by providing email and password
      parameters:
      - description: user login details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.LoginUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Login a user
      tags:
      - Authentication
  /user/signup:
    post:
      consumes:
      - application/json
      description: Register an new user by providing email, password, first name,
        last name
      parameters:
      - description: user register details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - Authentication
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"